service/requirements.txt
fastapi==0.111.0
pydantic
pydantic-settings
uvicorn
psycopg2-binary
sqlalchemy==2.0.31
uuid==1.30